+++ handlers/settings.go
@@ -482,7 +482,7 @@ func (h settings) exportDownload(w http.ResponseWriter, r *http.Request) error {
 	fp, err := os.Open(export.Path)
 	if err != nil {
 		if os.IsNotExist(err) {
-			zhttp.FlashError(w, "It looks like there is no export yet.")
+			zhttp.FlashError(w, "It looks like there is no export yet or the export has expired")
 			return zhttp.SeeOther(w, "/settings/export")
 		}